+++ ttnn/cpp/ttnn/operations/normalization/groupnorm/device/groupnorm_program_utils.hpp
@@ -1,0 +1,112 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <vector>
+#include <cstdint>
+#include <limits>
+#include <algorithm>
+
+#include <tt-metalium/core_coord.hpp>
+#include <tt-metalium/constants.hpp>
+
+namespace ttnn::operations::normalization::group_norm {
+
+enum class GroupNormMode : uint32_t { LEGACY = 0, WELFORD_NATIVE = 1, WELFORD_RECIPROCALS = 2 };
+
+inline int get_max_subblock(uint32_t n, uint32_t max_subblock_w) {
+    if (n <= max_subblock_w) {
+        return n;
+    }
+
+    for (int quotient = max_subblock_w; quotient > 1; --quotient) {
+        if (n % quotient == 0) {
+            return quotient;
+        }
+    }
+    return 1;
+}
+
+inline bool is_rectangle_grid(const std::vector<CoreCoord>& core_coords) {
+    if (core_coords.empty()) {
+        return true;
+    }
+
+    int min_x = std::numeric_limits<int>::max();
+    int max_x = std::numeric_limits<int>::min();
+    int min_y = std::numeric_limits<int>::max();
+    int max_y = std::numeric_limits<int>::min();
+
+    for (const auto& coord : core_coords) {
+        min_x = std::min(min_x, static_cast<int>(coord.x));
+        max_x = std::max(max_x, static_cast<int>(coord.x));
+        min_y = std::min(min_y, static_cast<int>(coord.y));
+        max_y = std::max(max_y, static_cast<int>(coord.y));
+    }
+
+    return ((max_x - min_x + 1) * (max_y - min_y + 1)) == static_cast<int>(core_coords.size());
+}
+
+inline void split_and_form_rectangle_grids(
+    std::vector<CoreCoord>& group,
+    std::vector<CoreCoord>& mcast_group_first,
+    std::vector<CoreCoord>& mcast_group_mid,
+    std::vector<CoreCoord>& mcast_group_last) {
+    size_t remove_front = 0;
+    size_t remove_back = 0;
+    size_t min_total_removal = group.size();
+
+    for (size_t front = 0; front <= group.size(); ++front) {
+        for (size_t back = 0; front + back <= group.size(); ++back) {
+            if (is_rectangle_grid(std::vector<CoreCoord>(group.begin() + front, group.end() - back))) {
+                size_t total_removal = front + back;
+                if (total_removal < min_total_removal) {
+                    min_total_removal = total_removal;
+                    remove_front = front;
+                    remove_back = back;
+                }
+            }
+        }
+    }
+
+    // Pop and push the front outliers
+    for (size_t i = 0; i < remove_front; ++i) {
+        mcast_group_first.push_back(mcast_group_mid.front());
+        mcast_group_mid.erase(mcast_group_mid.begin());
+    }
+
+    // Pop and push the back outliers
+    for (size_t i = 0; i < remove_back; ++i) {
+        mcast_group_last.push_back(mcast_group_mid.back());
+        mcast_group_mid.pop_back();
+    }
+}
+
+inline std::pair<uint32_t, uint32_t> find_max_tile_span(uint32_t W, uint32_t group_size) {
+    uint32_t current_position = 0;
+    uint32_t max_tile_span = 0;
+    uint32_t num_groups_before_start_again_at_tile_beginning = static_cast<uint32_t>(-1);
+    bool calc_num_groups_before_start_again_at_tile_beginning = true;
+
+    while (current_position < W) {
+        uint32_t group_end = current_position + group_size;
+        uint32_t start_tile = current_position / tt::constants::TILE_WIDTH;
+        uint32_t end_tile = (group_end - 1) / tt::constants::TILE_WIDTH;
+        uint32_t current_tile_span = end_tile - start_tile + 1;
+
+        max_tile_span = std::max(max_tile_span, current_tile_span);
+
+        current_position = group_end;
+
+        if (current_position % tt::constants::TILE_WIDTH == 0 && calc_num_groups_before_start_again_at_tile_beginning) {
+            num_groups_before_start_again_at_tile_beginning = current_position / group_size;
+            calc_num_groups_before_start_again_at_tile_beginning = false;
+        }
+    }
+
+    return {max_tile_span, num_groups_before_start_again_at_tile_beginning};
+}
+
+}  // namespace ttnn::operations::normalization::group_norm
