+++ ttnn/cpp/ttnn/operations/normalization/groupnorm/groupnorm_input_mask.hpp
@@ -6,7 +6,7 @@
 
 #include <cstdint>
 #include <vector>
-#include "device/groupnorm_types.hpp"
+#include "ttnn/tensor/tensor.hpp"
 
 namespace ttnn::operations::normalization {
 
@@ -14,9 +14,15 @@ namespace ttnn::operations::normalization {
 // Create 4D mask [1, num_groups, 32, 32*block_wt] used by group norm.
 // block_wt is computed from worst-case tile span across groups.
 // num_cores_across_channel splits groups evenly across cores (must divide num_groups).
-ttnn::Tensor create_group_norm_input_mask(int64_t num_channel, int64_t num_groups,
-      int64_t num_cores_across_channel, DataType data_type = DataType::BFLOAT16);
+ttnn::Tensor create_group_norm_input_mask(
+    int64_t num_channel,
+    int64_t num_groups,
+    int64_t num_cores_across_channel,
+    tt::tt_metal::DataType data_type = tt::tt_metal::DataType::BFLOAT16);
 
 ttnn::Tensor create_group_norm_input_negative_mask(
-    int64_t num_channel, int64_t num_groups, int64_t num_cores_across_channel, DataType data_type = DataType::BFLOAT16);
+    int64_t num_channel,
+    int64_t num_groups,
+    int64_t num_cores_across_channel,
+    tt::tt_metal::DataType data_type = tt::tt_metal::DataType::BFLOAT16);
 }  // namespace normalization
